third_party/WebKit/LayoutTests/http/tests/inspector/timeline-xhr-event-expected.txt

CONSOLE WARNING: The XMLHttpRequest progress event property 'totalSize' is deprecated. Please use 'total' instead.
CONSOLE WARNING: The XMLHttpRequest progress event property 'position' is deprecated. Please use 'loaded' instead.
Tests the Timeline events for XMLHttpReqeust

XHRReadyStateChange Properties:
{
    children : <object>
    data : {
        readyState : 2
        url : .../inspector/network/resources/resource.php
    }
    endTime : <number>
    frameId : <string>
    startTime : <number>
    thread : <string>
    type : "XHRReadyStateChange"
}
XHRReadyStateChange Properties:
{
    children : <object>
    data : {
        readyState : 3
        url : .../inspector/network/resources/resource.php
    }
    endTime : <number>
    frameId : <string>
    startTime : <number>
    thread : <string>
    type : "XHRReadyStateChange"
}
XHRReadyStateChange Properties:
{
    children : <object>
    data : {
        readyState : 4
        url : .../inspector/network/resources/resource.php
    }
    endTime : <number>
    frameId : <string>
    startTime : <number>
    thread : <string>
    type : "XHRReadyStateChange"
}
XHRLoad Properties:
{
    children : <object>
    data : {
        url : .../inspector/network/resources/resource.php
    }
    endTime : <number>
    frameId : <string>
    startTime : <number>
    thread : <string>
    type : "XHRLoad"
}
